requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
docopt==0.6.2
idna==3.3
python-dotenv==0.19.2
requests==2.26.0
tk==0.1.0
urllib3==1.26.7
yarg==0.1.9
